kremlib/c: reorganize headers for uint128 support
Suggested by Jonathan Protzenko in:
https://github.com/FStarLang/kremlin/pull/173#issuecomment-594734879
