+++ include/kremlin/internal/types.h
@@ -49,19 +49,20 @@ typedef const char *Prims_string;
 
 /* The great static header headache. */
 
+#define IS_MSVC64 (defined(_MSC_VER) && defined(_M_X64) && !defined(__clang__))
+#define HAS_INT128 ((defined(__x86_64__) || defined(__x86_64) || \
+                     defined(__aarch64__) || \
+                     (defined(__powerpc64__) && defined(__LITTLE_ENDIAN__))) || \
+                    (defined(_MSC_VER) && !defined(_M_X64) && defined(__clang__)))
+
 /* The uint128 type is a special case since we offer several implementations of
  * it, depending on the compiler and whether the user wants the verified
  * implementation or not. */
-#if !defined(KRML_VERIFIED_UINT128) && defined(_MSC_VER) && defined(_M_X64) && \
-    !defined(__clang__)
+#if !defined(KRML_VERIFIED_UINT128) && IS_MSVC64
 #  include <emmintrin.h>
 typedef __m128i FStar_UInt128_uint128;
-#elif !defined(KRML_VERIFIED_UINT128) && !defined(_MSC_VER) && \
-      (defined(__x86_64__) || defined(__x86_64) || defined(__aarch64__) || \
-      (defined(__powerpc64__) && defined(__LITTLE_ENDIAN__)))
+#elif !defined(KRML_VERIFIED_UINT128) && HAS_INT128
 typedef unsigned __int128 FStar_UInt128_uint128;
-#elif !defined(KRML_VERIFIED_UINT128) && defined(_MSC_VER) && defined(__clang__)
-typedef __uint128_t FStar_UInt128_uint128;
 #else
 typedef struct FStar_UInt128_uint128_s {
   uint64_t low;
@@ -75,15 +76,13 @@ typedef FStar_UInt128_uint128 FStar_UInt128_t, uint128_t;
 
 #include "kremlin/lowstar_endianness.h"
 
-/* This one is always included, because it defines C.Endianness functions too. */
-#if !defined(_MSC_VER) || defined(__clang__)
+#if !defined(KRML_VERIFIED_UINT128) && HAS_INT128
 #include "fstar_uint128_gcc64.h"
-#endif
-
-#if !defined(KRML_VERIFIED_UINT128) && defined(_MSC_VER) && !defined(__clang__)
+#if !defined(KRML_VERIFIED_UINT128) && IS_MSVC64
 #include "fstar_uint128_msvc.h"
 #elif defined(KRML_VERIFIED_UINT128)
 #include "FStar_UInt128_Verified.h"
+#include "fstar_uint128_struct_endianness.h"
 #endif
