+++ kremlib/c/fstar_uint128_gcc64.h
@@ -24,10 +24,6 @@
 #include "FStar_UInt_8_16_32_64.h"
 #include "LowStar_Endianness.h"
 
-#if !defined(KRML_VERIFIED_UINT128) && (!defined(_MSC_VER) || defined(__clang__)) && \
-      (defined(__x86_64__) || defined(__x86_64) || defined(__aarch64__) || \
-      (defined(__powerpc64__) && defined(__LITTLE_ENDIAN__)))
-
 /* GCC + using native unsigned __int128 support */
 
 inline static uint128_t load128_le(uint8_t *b) {
@@ -162,67 +158,3 @@ inline static bool FStar_UInt128_lte(uint128_t x, uint128_t y) {
 inline static uint128_t FStar_UInt128_mul32(uint64_t x, uint32_t y) {
   return (uint128_t) x * (uint128_t) y;
 }
-
-#elif !defined(_MSC_VER) && defined(KRML_VERIFIED_UINT128)
-
-/* Verified uint128 implementation. */
-
-/* Access 64-bit fields within the int128. */
-#define HIGH64_OF(x) ((x)->high)
-#define LOW64_OF(x)  ((x)->low)
-
-/* A series of definitions written using pointers. */
-
-inline static void load128_le_(uint8_t *b, uint128_t *r) {
-  LOW64_OF(r) = load64_le(b);
-  HIGH64_OF(r) = load64_le(b + 8);
-}
-
-inline static void store128_le_(uint8_t *b, uint128_t *n) {
-  store64_le(b, LOW64_OF(n));
-  store64_le(b + 8, HIGH64_OF(n));
-}
-
-inline static void load128_be_(uint8_t *b, uint128_t *r) {
-  HIGH64_OF(r) = load64_be(b);
-  LOW64_OF(r) = load64_be(b + 8);
-}
-
-inline static void store128_be_(uint8_t *b, uint128_t *n) {
-  store64_be(b, HIGH64_OF(n));
-  store64_be(b + 8, LOW64_OF(n));
-}
-
-#    ifndef KRML_NOSTRUCT_PASSING
-
-inline static uint128_t load128_le(uint8_t *b) {
-  uint128_t r;
-  load128_le_(b, &r);
-  return r;
-}
-
-inline static void store128_le(uint8_t *b, uint128_t n) {
-  store128_le_(b, &n);
-}
-
-inline static uint128_t load128_be(uint8_t *b) {
-  uint128_t r;
-  load128_be_(b, &r);
-  return r;
-}
-
-inline static void store128_be(uint8_t *b, uint128_t n) {
-  store128_be_(b, &n);
-}
-
-#    else /* !defined(KRML_STRUCT_PASSING) */
-
-#      define print128 print128_
-#      define load128_le load128_le_
-#      define store128_le store128_le_
-#      define load128_be load128_be_
-#      define store128_be store128_be_
-
-#    endif /* KRML_STRUCT_PASSING */
-
-#endif
